feat: hearts page suspens, errorbound 추가
